+++ docs/assets/vendor-react-B61y0MB6-0.3.7.js
@@ -1,4 +1,4 @@
-import{g as Ua,a as $a}from"./vendor-DQhf2rhQ-0.3.6.js";var Pi={exports:{}},Ar={},zi={exports:{}},N={};/**
+import{g as Ua,a as $a}from"./vendor-DQhf2rhQ-0.3.7.js";var Pi={exports:{}},Ar={},zi={exports:{}},N={};/**
  * @license React
  * react.production.min.js
  *